[Mod]: Introduce serializer_extra
serializer_extra is a dictionary of extra keyword arguments to pass to
serializer instantiation. To use it, one would need to add the
dictionary into the kwargs argument received at the class-based view methods
that process requests. It is, however, optional. GenericAPIView and all
of the mixins are modfied to reflect this change.